+++ test/unit/auth/user-record.spec.ts
@@ -20,14 +20,13 @@ import * as chaiAsPromised from 'chai-as-promised';
 
 import { deepCopy } from '../../../src/utils/deep-copy';
 import {
-  UserInfo, UserMetadata, UserRecord,
-  PhoneMultiFactorInfo, MultiFactorInfo,
+  UserInfo, UserMetadata, UserRecord, PhoneMultiFactorInfo
 } from '../../../src/auth/user-record';
 import { 
   GetAccountInfoUserResponse, ProviderUserInfoResponse,
   MultiFactorInfoResponse
 } from '../../../src/auth/user-record-internal';
-import { MultiFactor } from '../../../src/auth/user-record-internal';
+import { MultiFactor, initMultiFactorInfo } from '../../../src/auth/user-record-internal';
 
 
 chai.should();
@@ -390,11 +389,11 @@ describe('MultiFactorInfo', () => {
 
   describe('initMultiFactorInfo', () => {
     it('should return expected PhoneMultiFactorInfo', () => {
-      expect(MultiFactorInfo.initMultiFactorInfo(serverResponse)).to.deep.equal(phoneMultiFactorInfo);
+      expect(initMultiFactorInfo(serverResponse)).to.deep.equal(phoneMultiFactorInfo);
     });
 
     it('should return null for invalid MultiFactorInfo', () => {
-      expect(MultiFactorInfo.initMultiFactorInfo(undefined as any)).to.be.null;
+      expect(initMultiFactorInfo(undefined as any)).to.be.null;
     });
   });
 });
